reorganise the code
