+++ src/super_gradients/common/crash_handler/crash_tips.py
@@ -1,8 +1,15 @@
+import os
+import atexit
 import re
 from typing import Union, Tuple, List, Type
 from types import TracebackType
 
 from super_gradients.common.crash_handler.utils import indent_string, fmt_txt, json_str_to_dict
+from super_gradients.common.crash_handler.exception import ExceptionInfo
+from super_gradients.common.abstractions.abstract_logger import get_logger
+
+
+logger = get_logger(__name__)
 
 
 class CrashTip:
@@ -173,3 +180,20 @@ def get_relevant_crash_tip_message(exc_type: type, exc_value: Exception, exc_tra
         if crash_tip.is_relevant(exc_type, exc_value, exc_traceback):
             return crash_tip.get_message(exc_type, exc_value, exc_traceback)
     return None
+
+
+def crash_tip_handler():
+    """Display a crash tip if an error was raised"""
+    crash_tip_message = ExceptionInfo.get_crash_tip_message()
+    if crash_tip_message:
+        print(crash_tip_message)
+
+
+def setup_crash_tips() -> bool:
+    if os.getenv("CRASH_HANDLER", "TRUE") != "False":
+        logger.info("Crash tips is enabled. You can set your environment variable to CRASH_HANDLER=FALSE to disable it")
+        atexit.register(crash_tip_handler)
+        return True
+    else:
+        logger.info("Crash tips is disabled. You can set your environment variable to CRASH_HANDLER=TRUE to enable it")
+        return False
